refactor(loom): tighten layering boundaries
